fix: Split the data treatment into multiple files Other changes: - Add `switchIndex` for the Pokemon Trainer, Squirtle, Ivysaur and Charizard ; - Change `gameSeries` value from `Pokemon` to `Pokémon` (same change to   the Pokemon Trainer name) ; - Replace import method by a returning type (array).
